This adds a few new optional inputs to Project install apis.
`project.resolveEverything` accepts a custom resolver, it will function as before with out this option.
`project.install` accepts a custom resolver. It is used for `resolveEverything`, `fetchEverything`, and `linkEverything`
`project.linkEverything` accepts a custom fetcher, it will function as before with out this option.

`linkEverything` was changed because if install used a custom fetcher it would fail an install during the linking stage if the fetcher changed functionally. This allows for the same fetcher that was using during install to be used during the linking stage.

`install` and `resolveEverything` were changed to allow new functionally in the api. The change allows a plugin to provide custom resolver logic. This complements the custom fetcher in cases where the fetcher would allow for resolutions not supported by the ones provided within configuration.