refine targets for header hyperlinks
When I hover over the banner that says,
     "For the 1.x doc", I get a hyperlink, and so I clicked that
     expecting to be taken to v1 documentation.

Unfortunately that part of the sentence is linked to the berry
repository. When I arrived there, I assumed that I just needed to switch
to a v1.x branch or tag, but no such things exist on that repository.

After bouncing around a bit more, the second time I clicked on the
correct portion of the banner and was taken to production documentation.

I'm hoping that this change can help others avoid confusion.