fix(patch-commit): diffFolders returned silently when 'git' failed
Fixes #2331

The `diffFolders` function runs `git` and disregarded the return status `code` and `stderr`, which means it resolved with an empty value of `stdout`.

The default `git` that's included with OS X does not support one of the arguments passed to it. Example:

```sh
❯ git diff --src-prefix=a/ --dst-prefix=b/ --ignore-cr-at-eol --full-index --no-index /private/.../xfs-adf9c270 /private/.../xfs-e4548122
fatal: invalid diff option/value: --ignore-cr-at-eol
```

Berry thinks an empty string is the result of the diff, prints it to stdout and incorrectly returns status 0.