fix(core): tarball extraction incorrectly stripping components/parts
Yarn 1.x always strips the first component away. Similarly Yarn berry
attempts this but fails due to some normalization logic applied to deal
with its portable paths. The result is that before stripping the first component,
Yarn berry accidentally normalizes paths, throwing off the component stripping of
tarball entries.

e.g. if a tarball (that works in Yarn 1.x) has entries like:

```
./package.json
./index.js
```

the entries above will be discarded even though `.` is the first component/part. The
issue is that Yarn berry normalizes the paths to just e.g. `package.json`, causing the
entry parts to be smaller/equal to the strip count, and the entry being discarded.

The fix is to only apply the stripping logic on the unmodified path, except for the
path being converted to forward slashes.

Fixes #4070.