[plugin-npm-cli]: Add ability to exclude packages, or ignore specific advisories in `yarn npm audit` (#4356)
* [plugin-npm-cli]: Add ability to exclude packages from `yarn npm audit`

This patch adds a `--exclude` flag to the `yarn npm audit` command in the
`nmp-cli` plugin. This flag can be passed multiple times, and any package
listed will be removed from the list of packages audited.

* [plugin-npm-cli] Add ability to ignore advisories in `yarn npm audit`

This patch adds a `--ignore` flag to `yarn npm audit`, which is an array
of ID's to ignore from the audit report.

In addition, the ID is presented in the tree output (as well as the JSON).

* Version bump

* chore: Fix types

* [plugin-npm-cli] Add configuration options for --exclude and --ignore

Adds configuration options to specify packages to exclude from `yarn npm audit`
and to specify advisories to ignore from the results.

* Update audit.ts

* Update audit.ts

* [plugin-npm-cli] Update docs

* [plugin-npm-cli] Add support for glob patterns in --exclude and --ignore

* [plugin-npm-cli] Add some unit tests and stubs for integration tests

Co-authored-by: Maël Nison <nison.mael@gmail.com>