Implements .env file support (#5531)
**What's the problem this PR addresses?**

A common need is to provide environment values into the environment via
`.env` files. There's been a couple of issues and attempts at
implementation already, which all were decently upvoted. I myself could
have used it once or twice 😄

Props to @jj811208 for his implementation in #4835 - I wanted to make
the configuration a little more generic (allowing to have multiple
environment files, and to possibly disable it altogether), but it was a
appreciated start.

Fixes #4718 
Closes #4835 (Supercedes it)

**How did you fix it?**

A new setting, `injectEnvironmentFiles`, lets you define files that Yarn
will load and inject into all scripts. It only affects subprocesses -
Yarn itself still uses `process.env` for its checks, so you can't for
example set `YARN_*` values and expect them to be applied to the current
process (use the yarnrc file for that instead).

The `injectEnvironmentFiles` setting has a few properties:

- It defaults to `.env`
- Nothing will be injected if it's set to an empty array or null
- The paths inside may be suffixed by `?` - in that case, Yarn won't
throw if the file doesn't exist

The idea with this last property is to allow for simple user
configuration (imagine, with the example below, that the project also
has a gitignore with `.env.*`):

```
injectEnvironmentFiles:
  - .env
  - .env.${USER}?
```

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.