Cache tolerance for old archives (#5564)
**What's the problem this PR addresses?**

Each time we make a change to the libzip or zlib implementation, all
generated archives typically see the compression output changed, with no
notable changes on the actual content of the files. When the local cache
is used this is reflected by git marking all cache files as changed, and
requiring to push them anew. This is a very heavy process, especially
since the compression prevents git from efficiently diffing the files.
The situation is a little better with compression=0 (the new default),
but not all projects can use it right away.

**How did you fix it?**

This diff makes Yarn tolerant to archives generated by older versions,
as long as they have been generated by at least the cache version N
(called the "checkpoint version"). The checksum is still validated.

This logic isn't enabled by default, by fear of adverse interactions
with particular setups (in particular `--check-cache`, which is
currently recommended for open-source projects using zero-installs;
probably not many, but better to be careful so close from the next major
release).

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.