Refactors the CLI spawn (#5600)
**What's the problem this PR addresses?**

While working on the new website, I find myself in a situation where I
need to access Yarn's `cli` instance (I want to use it to tokenize the
examples on the website). This is difficult at the moment, as all parts
of the CLI initialization are entangled together.

**How did you fix it?**

I extracted our initialization steps into separate functions, each
called in sequence. The one breaking change I made is that `--cwd` is
now only accepted as very first argument of the command line (this is
needed to get rid of the recursive execution which caused various
issues) - if it isn't, Yarn will throw an error explaining the problem.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.