refactor!: remove `ignoreCwd` (#5602)
**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

Follow-up to https://github.com/yarnpkg/berry/pull/5600.

This PR removes `ignoreCwd` because the option was only a broken
workaround that doesn't have a use anymore and even still leads to bugs
(try running `yarn --cwd=packages exec pwd` on `master`).

**How did you fix it?**
<!-- A detailed description of your implementation. -->

Removed it as an option but still kept it as an ignored env variable as
it's set by the parent process which could be anything (including a Yarn
version that still sets it).

Tests will come in a future PR.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [X] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [X] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [X] I will check that all automated PR checks pass before the PR gets
reviewed.