Extracts the repository plugin hooks into their own file (#5622)
**What's the problem this PR addresses?**

The repository is configured to compile files dynamically, but this only
works for the CLI - we currently have no way to require the plugin
configuration without executing Yarn. I will need that for #5587, which
will use the CLI instance to tokenize the CLI commands but without
actually executing them.

**How did you fix it?**

The script hooking the repository into the plugin configuration is
extracted in a standalone script.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.