Updates Clipanion to 4.0.0-rc.1 (#5624)
**What's the problem this PR addresses?**

I will need command tokenization, which is only available in 4.0.0-rc.1
since it removes the old suggestion engine.

While this version is technically unstable, it's developed close enough
from Yarn itself that we can ensure minimal disruption, so I feel
comfortable using it before it's truly ready to be shipped (at least
more than the alternative, which would be to have to release many majors
on Clipanion each time we need to make a fairly small change for the
sake of Yarn).

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.