refactor!: remove `lockfileFilename` (#5604)
**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

`lockfileFilename` is a configuration option that provides little value
and overcomplicates other parts of the codebase.

Many third-party tools aren't even aware of it and even we have parts of
the codebase that completely ignore it and use `yarn.lock`.

It also complicates project detection, as we first need to instantiate
the configuration before being able to walk up the filesystem to find
the project root.

**How did you fix it?**
<!-- A detailed description of your implementation. -->

Removed it.

`Configuration.findProjectCwd` could also use a different signature but
I'll leave it for a different PR since I intend to refactor it anyways.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [X] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [X] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [X] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Maël Nison <nison.mael@gmail.com>