Make simplifyRanges handle wildcards
This commit is a fix for (#5366)

simplifyRanges now ignores the wildcards in the range list that is given as
parameter when the function is called. Since a wildcard '*' is completely
permissive and does not narrow the range of versions that is supposed to
be simplified, the wildcard is removed from the list of ranges to simplify.