simplifyRanges now handles wildcards (#5679)
**What's the problem this PR addresses?**
When yarn used the new functionality that was introduced in
155ff9cc4976a214fc221a0fe7566da12ab674a2 a crash occured when a wildcard
'*' was to considered in a version range that should be simplified.

This fixes issue #5633
...

**How did you fix it?**
simplifyRanges now ignores the wildcards in the range list that is given
as parameter when the function is called. Since a wildcard '*' is
completely permissive and does not narrow the range of versions that is
supposed to be simplified, the wildcard is removed from the list of
ranges to simplify.


**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Ulrik Sjölin <ulrik.sjolin@sensative.com>
Co-authored-by: Maël Nison <nison.mael@gmail.com>