Fixes global cache (#5739)
**What's the problem this PR addresses?**

The cache fix merged in https://github.com/yarnpkg/berry/pull/5730 had a
bug that caused the global cache to be skipped when a package had no
checksum - which happens when there are no lockfile in the project. This
was detected by the perf regression visible on the dashboard:

<img width="1381" alt="image"
src="https://github.com/yarnpkg/berry/assets/1037931/11aed9d5-7946-418d-806a-6df78aee4734">

**How did you fix it?**

I fixed the condition (it was "use the cache if there's a checksum and
it's compatible", it should have been "use the cache if there's no
checksum, or it's a compatible one") and added tests to avoid further
regressions.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.