Reworks the worker pool to be more configurable (#5764)
**What's the problem this PR addresses?**

This PR extends @anuragkalia's work in #5636 (it starts from the same
commits and builds upon them).

- The original implementation was keeping a singleton worker pool - I
felt this was problematic since we may enter situations (even if only in
tests) where multiple calls to `convertToZip` would want to use
different pools.

- Converting an archive without using a worker was only possible when
the concurrency was completely disabled. This was my idea, but
retrospectively it felt better to me to have two settings: one for the
concurrency, and another to define whether the concurrency is enabled or
not.

- Passing all those settings from the various fetchers would have been
unwieldly.

**How did you fix it?**

- I created a `TaskPool` interface; `WorkerPool` implements it (using
workers, same implementation as before), but so does `AsyncPool` (a new
implementation that simply forwards the call to the `pLimit` limiter).

- I feel like this also addresses @merceyz's comment regarding the
setting name - it's now called `taskPoolConcurrency`, which doesn't
directly refers to workers.

- A `getConfigurationWorker` function returns the proper task pool for
the given configuration object. To make that possible, WeakMap instances
can now be used as storage argument for the `get*WithDefault` family of
functions.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Anurag Kalia <akalia@atlassian.com>