Fixes bugs in yarn npm audit (#5833)
**What's the problem this PR addresses?**

The rewrite of `yarn npm audit` broke a couple of things in the advisory
reporting.

Fixes #5824 - The `--exclude` filter wasn't applied to deprecations
Fixes #5830 - The `--json` output was printing the raw registry output,
not the filtered tree

**How did you fix it?**

Fixed the deprecation iterations (we were iterating the unfiltered
data).

Fixed the json reporting to instead print the same tree as the one we
would have displayed outside of `--json`. It's _technically_ a breaking
change, but it was supposed to be like that so it's probably fine to
treat it as a 7.0.1 bugfix.

Added tests to prevent regressions.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Wojciech Maj <kontakt@wojtekmaj.pl>