Adds tests for the pack list generation (#5873)
**What's the problem this PR addresses?**

#5872 made me notice that we have few tests covering how the `*` / `**`
patterns are interpreted.

**How did you fix it?**

This diff fixes that by adding a couple of tests, along with comments on
the few places where we seem to diverge from npm. It doesn't try to
change the pack list generation, just to increase its coverage and make
it easier to add more tests in the future.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.