fix: remove unused --save option for set resolution command
The option was never implemented in berry, but it's mentioned in the docs as well
as in the CLI output.

See #2202

I don't have the expertise to actually implement the flag, so the second best thing to do here
is to remove it, so that people are no longer mislead.