fix: remove unused --save option for set resolution command (#5868)
The option was never implemented in berry, but it's mentioned in the
docs as well as in the CLI output.

I don't have the expertise to actually implement the flag, so the second
best thing to do here is to remove it, so that people are no longer
mislead.

**What's the problem this PR addresses?**

Mitigates https://github.com/yarnpkg/berry/issues/2202 (actual
resolution should be implementing the flag).

**How did you fix it?**

Removed unnecessary flag.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Clement Yan <clemyan@users.noreply.github.com>