Adds a warning when the hardened mode is enabled (#5969)
**What's the problem this PR addresses?**

It's currently difficult to spot that the hardened mode is enabled;
people upgrading to Yarn 4 will think that the resolution is slower due
to a bug, without noticing that the slowdown is caused by the extra
check on the remote registry.

**How did you fix it?**

This diff adds a warning prior to the install when the hardened mode is
enabled.

I'd also like to be smarter in a subsequent diff and skip the hardened
mode when the user is the repository admin, but that will be for a
separate PR.

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.