Add API Reference to website (#5922)
**What's the problem this PR addresses?**
<!-- Describe the rationale of your PR. -->
<!-- Link all issues that it closes. (Closes/Resolves #xxxx.) -->

The API reference is not available in new website

**How did you fix it?**
<!-- A detailed description of your implementation. -->

This is a WIP to bring the API reference to the new website via
`docusaurus-plugin-typedoc-api`.

The current state is has some kinks and issues to work out. The purpose
of this draft PR is to track and discuss. I'll sort those out over time
but if anyone wants to take over then be my guest.

- ~~Seems like `docusaurus-plugin-typedoc-api` need to use the instance
from `@docusaurus/theme-common` and `@docusaurus/plugin-content-docs`
used by the main doc, which were hidden inside
`@docusaurus/preset-classic`. I added package extensions to resolve them
but that seems brittle.~~
- ~~Docusaurus's MDX loader resolves `react` and `@mdx-js/react` from
the loaded Markdown file. To include READMEs into the API reference we
need every workspace to depend on `react` and `@mdx-js/react` which is
not very reasonable. For now I use `pnpFallbackMode: all` so they
fallback to the `react` and `@mdx-js/react` in the top-level but again
seems brittle.~~
- ~~Only resolve one entrypoint for each workspace. Tried multiple
entries and some pages broke. Maybe try again later?~~
- ~~Some stuff have duplicate source links due to virtual paths. Need to
deduplicate~~
- ~~Functions receive dedicated pages but the header is just "Callable".
Ideally I want them to be just in the index page like variables~~

**Checklist**
<!--- Don't worry if you miss something, chores are automatically
tested. -->
<!--- This checklist exists to help you remember doing the chores when
you submit a PR. -->
<!--- Put an `x` in all the boxes that apply. -->
- [x] I have read the [Contributing
Guide](https://yarnpkg.com/advanced/contributing).

<!-- See
https://yarnpkg.com/advanced/contributing#preparing-your-pr-to-be-released
for more details. -->
<!-- Check with `yarn version check` and fix with `yarn version check
-i` -->
- [x] I have set the packages that need to be released for my changes to
be effective.

<!-- The "Testing chores" workflow validates that your PR follows our
guidelines. -->
<!-- If it doesn't pass, click on it to see details as to what your PR
might be missing. -->
- [x] I will check that all automated PR checks pass before the PR gets
reviewed.

---------

Co-authored-by: Maël Nison <nison.mael@gmail.com>